feat: CLIN-1737 Added new tracks to IGV
